requirements.txt
flask
transformers
torch
torchvision
opencv-python
pillow
timm